+++ packages/@pufflig/ps-chains/src/engines/dataflow/updateNodeInput.ts
@@ -1,6 +1,7 @@
 import { ParamValue } from "@pufflig/ps-types";
 import { Flow, NodeState, RunOptions } from "../../types";
 import { applyDefaultInputs } from "./utils/utils";
+import { identity } from "./constants";
 
 /**
  * This function updates the input state of a node without executing the chain.
@@ -20,6 +21,7 @@ export async function updateNodeInput(
   const flowState: Record<string, NodeState> = { ...flow.state };
   const nodeConfig = flow.definition.nodes[nodeId];
   const nodeDefinition = flow.nodeTypes[nodeConfig?.type];
+  const mapInput = nodeDefinition?.mapInput || identity;
 
   if (!nodeConfig || !nodeDefinition) {
     throw new Error(`Definition for node ${nodeId} not found`);
@@ -28,7 +30,7 @@ export async function updateNodeInput(
   // parse the input
   const nodeState = flowState[nodeId];
   const prevInput = nodeState?.input;
-  const parsedInput = await nodeDefinition.parseInput(input, prevInput);
+  const parsedInput = await mapInput(input, prevInput);
 
   // compile the new state of the node
   const newInput = { ...applyDefaultInputs(prevInput, nodeDefinition), ...parsedInput };