+++ packages/@pufflig/ps-nodes/src/adapters/custom_api/custom_api_chat.ts
@@ -13,8 +13,11 @@ export interface CustomAPIChatOutput {
   chat: Chat;
 }
 
-export const customApiChat: Node = {
+export const customApiChat: Node<CustomAPIChatInput, CustomAPIChatOutput> = {
   ...nodes[customApiChatNodeType],
-  execute: async (i) => i,
-  parseInput: async (i) => i,
+  execute: async () => {
+    return {
+      chat: { messages: [] },
+    };
+  },
 };