+++ packages/@pufflig/ps-nodes/src/adapters/openai/openai_chat.ts
@@ -1,21 +1,21 @@
 import { nodes } from "@pufflig/ps-nodes-config";
-import { Chat, ChatMessage, ModelValue, Node } from "@pufflig/ps-types";
-import { Configuration, OpenAIApi, ChatCompletionRequestMessage } from "openai";
+import { Chat, ChatMessage, ModelValue, Node, ParamValueMap } from "@pufflig/ps-types";
+import { ChatCompletionRequestMessage, Configuration, OpenAIApi } from "openai";
 import { v4 as uuid } from "uuid";
 
 export const openaiChatNodeType = "adapter/openai_chat" as const;
 
-export interface OpenAIChatInput {
+export interface OpenAIChatInput extends ParamValueMap {
   api_key: string;
   chat: Chat;
   model: ModelValue;
 }
 
-export interface OpenAIChatOutput {
+export interface OpenAIChatOutput extends ParamValueMap {
   message: ChatMessage;
 }
 
-export const execute = async (input: OpenAIChatInput): Promise<OpenAIChatOutput> => {
+export const execute = async (input: OpenAIChatInput) => {
   const { chat, model, api_key } = input;
   const { modelId, parameters } = model;
 
@@ -42,8 +42,7 @@ export const execute = async (input: OpenAIChatInput): Promise<OpenAIChatOutput>
   };
 };
 
-export const openaiChat: Node = {
+export const openaiChat: Node<OpenAIChatInput, OpenAIChatOutput> = {
   ...nodes[openaiChatNodeType],
   execute,
-  parseInput: async (i) => i,
 };