+++ packages/@pufflig/ps-nodes/src/data/text/text.ts
@@ -3,6 +3,4 @@ import { Node } from "@pufflig/ps-types";
 
 export const textNode: Node = {
   ...nodes[nodeTypes.textNodeType],
-  execute: async (i) => i,
-  parseInput: async (i) => i,
 };